Refactored LoanTransactionsLogic
