+++ app/src/main/java/com/ivy/wallet/logic/loantrasactions/LoanTransactionsCore.kt
@@ -1,0 +1,316 @@
+package com.ivy.wallet.logic.loantrasactions
+
+import androidx.compose.ui.graphics.toArgb
+import com.ivy.wallet.base.computationThread
+import com.ivy.wallet.base.ioThread
+import com.ivy.wallet.base.timeNowUTC
+import com.ivy.wallet.logic.currency.ExchangeRatesLogic
+import com.ivy.wallet.model.LoanType
+import com.ivy.wallet.model.TransactionType
+import com.ivy.wallet.model.entity.*
+import com.ivy.wallet.persistence.dao.*
+import com.ivy.wallet.sync.uploader.TransactionUploader
+import com.ivy.wallet.ui.IvyContext
+import com.ivy.wallet.ui.theme.components.IVY_COLOR_PICKER_COLORS_FREE
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import java.time.LocalDateTime
+import java.util.*
+
+class LoanTransactionsCore(
+    private val categoryDao: CategoryDao,
+    private val transactionUploader: TransactionUploader,
+    private val transactionDao: TransactionDao,
+    private val ivyContext: IvyContext,
+    private val loanRecordDao: LoanRecordDao,
+    private val loanDao: LoanDao,
+    private val settingsDao: SettingsDao,
+    private val accountsDao: AccountDao,
+    private val exchangeRatesLogic: ExchangeRatesLogic
+) {
+    private var baseCurrencyCode: String? = null
+
+    init {
+        CoroutineScope(Dispatchers.IO).launch {
+            baseCurrencyCode = baseCurrency()
+        }
+    }
+
+     suspend fun deleteAssociatedTransactions(
+        loanId: UUID? = null,
+        loanRecordId: UUID? = null
+    ) {
+        if (loanId == null && loanRecordId == null)
+            return
+
+        ioThread {
+            val transactions: List<Transaction?> =
+                if (loanId != null) transactionDao.findAllByLoanId(loanId = loanId) else
+                    listOf(transactionDao.findLoanRecordTransaction(loanRecordId!!))
+
+            transactions.forEach { trans ->
+                deleteTransaction(trans)
+            }
+        }
+    }
+
+     fun findAccount(
+        accounts: List<Account>,
+        accountId: UUID?,
+    ): Account? {
+        return accountId?.let { uuid ->
+            accounts.find { acc ->
+                acc.id == uuid
+            }
+        }
+    }
+
+     suspend fun baseCurrency(): String =
+        ioThread { baseCurrencyCode ?: settingsDao.findFirst().currency }
+
+
+     suspend fun updateAssociatedTransaction(
+        createTransaction: Boolean,
+        loanRecordId: UUID? = null,
+        loanId: UUID,
+        amount: Double,
+        loanType: LoanType,
+        selectedAccountId: UUID?,
+        title: String? = null,
+        category: Category? = null,
+        time: LocalDateTime? = null,
+        isLoanRecord: Boolean = false,
+        transaction: Transaction? = null,
+    ) {
+        if (isLoanRecord && loanRecordId == null)
+            return
+
+        if (createTransaction && transaction != null) {
+            createMainTransaction(
+                loanRecordId = loanRecordId,
+                loanId = loanId,
+                amount = amount,
+                loanType = loanType,
+                selectedAccountId = selectedAccountId,
+                title = title ?: transaction.title,
+                categoryId = category?.id ?: transaction.categoryId,
+                time = time ?: transaction.dateTime ?: timeNowUTC(),
+                isLoanRecord = isLoanRecord,
+                transaction = transaction
+            )
+        } else if (createTransaction && transaction == null) {
+            createMainTransaction(
+                loanRecordId = loanRecordId,
+                loanId = loanId,
+                amount = amount,
+                loanType = loanType,
+                selectedAccountId = selectedAccountId,
+                title = title,
+                categoryId = category?.id,
+                time = time ?: timeNowUTC(),
+                isLoanRecord = isLoanRecord,
+                transaction = transaction
+            )
+        } else {
+            deleteTransaction(transaction = transaction)
+        }
+    }
+
+    private suspend fun createMainTransaction(
+        loanRecordId: UUID? = null,
+        amount: Double,
+        loanType: LoanType,
+        loanId: UUID,
+        selectedAccountId: UUID?,
+        title: String? = null,
+        categoryId: UUID? = null,
+        time: LocalDateTime = timeNowUTC(),
+        isLoanRecord: Boolean = false,
+        transaction: Transaction? = null
+    ) {
+        if (selectedAccountId == null)
+            return
+
+        val transType = if (isLoanRecord)
+            if (loanType == LoanType.BORROW) TransactionType.EXPENSE else TransactionType.INCOME
+        else
+            if (loanType == LoanType.BORROW) TransactionType.INCOME else TransactionType.EXPENSE
+
+        val transCategoryId: UUID? = getCategoryId(existingCategoryId = categoryId)
+
+        val modifiedTransaction: Transaction = transaction?.copy(
+            loanId = loanId,
+            loanRecordId = if (isLoanRecord) loanRecordId else null,
+            amount = amount,
+            type = transType,
+            accountId = selectedAccountId,
+            title = title,
+            categoryId = transCategoryId,
+            dateTime = time
+        )
+            ?: Transaction(
+                accountId = selectedAccountId,
+                type = transType,
+                amount = amount,
+                dateTime = time,
+                categoryId = transCategoryId,
+                title = title,
+                loanId = loanId,
+                loanRecordId = if (isLoanRecord) loanRecordId else null
+            )
+
+        ioThread {
+            transactionDao.save(modifiedTransaction)
+        }
+    }
+
+    private suspend fun deleteTransaction(transaction: Transaction?) {
+        ioThread {
+            transaction?.let {
+                transactionDao.flagDeleted(it.id)
+            }
+
+            transaction?.let {
+                transactionUploader.delete(it.id)
+            }
+        }
+    }
+
+    private suspend fun getCategoryId(existingCategoryId: UUID? = null): UUID? {
+        if (existingCategoryId != null)
+            return existingCategoryId
+
+        val categoryList = ioThread {
+            categoryDao.findAll()
+        }
+
+        var addCategoryToDb = false
+
+        val loanCategory = categoryList.find { category ->
+            category.name.lowercase(Locale.ENGLISH).contains("loan")
+        } ?: if (ivyContext.isPremium || categoryList.size < 12) {
+            addCategoryToDb = true
+            Category(
+                "Loans",
+                color = IVY_COLOR_PICKER_COLORS_FREE[4].toArgb(),
+                icon = "loan"
+            )
+        } else null
+
+        if (addCategoryToDb)
+            ioThread {
+                loanCategory?.let {
+                    categoryDao.save(it)
+                }
+            }
+
+        return loanCategory?.id
+    }
+
+     suspend fun calculateLoanRecords(
+        newAccountId: UUID?,
+        loanId: UUID
+    ): List<LoanRecord> {
+        return ioThread {
+            val loanRecords =
+                loanRecordDao.findAllByLoanId(loanId).map { loanRecord ->
+                    val convertedAmount: Double? =
+                        computeConvertedAmount(
+                            newAmount = loanRecord.amount,
+                            loanAccountId = newAccountId,
+                            loanRecordId = loanRecord.id,
+                            loanRecordAccountId = loanRecord.accountId
+                        )
+                    loanRecord.copy(convertedAmount = convertedAmount)
+                }
+            loanRecords
+        }
+    }
+
+     suspend fun computeConvertedAmount(
+        newAmount: Double,
+        loanAccountId: UUID?,
+        loanRecordId: UUID,
+        loanRecordAccountId: UUID?,
+        reCalculateLoanAmount: Boolean = false,
+    ): Double? {
+        return computationThread {
+
+            val accounts = fetchAccounts()
+            val oldLoanRecord = fetchLoanRecord(loanRecordId)
+
+            val newLoanRecordCurrency =
+                loanRecordAccountId.fetchAssociatedCurrencyCode(accountsList = accounts)
+
+            val oldLoanRecordCurrency =
+                oldLoanRecord?.accountId.fetchAssociatedCurrencyCode(accountsList = accounts)
+
+            val loanCurrency = loanAccountId.fetchAssociatedCurrencyCode(accountsList = accounts)
+
+            val loanRecordCurrenciesChanged = oldLoanRecordCurrency != newLoanRecordCurrency
+
+            val newConverted: Double? = when {
+                newLoanRecordCurrency == loanCurrency -> {
+                    null
+                }
+
+                reCalculateLoanAmount || loanRecordCurrenciesChanged || oldLoanRecord?.convertedAmount == null || newLoanRecordCurrency != loanCurrency -> {
+                    ioThread {
+                        exchangeRatesLogic.convertAmount(
+                            baseCurrency = baseCurrency(),
+                            amount = newAmount,
+                            fromCurrency = newLoanRecordCurrency,
+                            toCurrency = loanCurrency
+                        )
+                    }
+                }
+
+                oldLoanRecord.amount != newAmount -> {
+                    newAmount * (oldLoanRecord.convertedAmount / oldLoanRecord.amount)
+                }
+
+                else -> {
+                    oldLoanRecord.convertedAmount
+                }
+            }
+            newConverted
+        }
+    }
+
+    private suspend fun UUID?.fetchAssociatedCurrencyCode(accountsList: List<Account>): String {
+        return findAccount(accountsList, this)?.currency ?: baseCurrency()
+    }
+
+     suspend fun fetchAccounts() = ioThread {
+        accountsDao.findAll()
+    }
+
+     suspend fun saveLoanRecords(loanRecords: List<LoanRecord>) = ioThread {
+        loanRecordDao.save(loanRecords)
+    }
+
+     suspend fun saveLoanRecords(loanRecord: LoanRecord) = ioThread {
+        loanRecordDao.save(loanRecord)
+    }
+
+     suspend fun saveLoan(loan: Loan) = ioThread {
+        loanDao.save(loan)
+    }
+
+     suspend fun fetchLoanRecord(loanRecordId: UUID) = ioThread {
+        loanRecordDao.findById(loanRecordId)
+    }
+
+     suspend fun fetchLoan(loanId: UUID) = ioThread {
+        loanDao.findById(loanId)
+    }
+
+    suspend fun fetchLoanRecordTransaction(loanRecordId: UUID?): Transaction? {
+        return loanRecordId?.let {
+            ioThread {
+                transactionDao.findLoanRecordTransaction(it)
+            }
+        }
+    }
+}
