+++ app/src/main/java/com/ivy/wallet/logic/loantrasactions/LoanTransactionsLogic.kt
@@ -1,0 +1,35 @@
+package com.ivy.wallet.logic.loantrasactions
+
+import com.ivy.wallet.base.computationThread
+import com.ivy.wallet.model.entity.Transaction
+
+data class LoanTransactionsLogic(
+    val Loan: LTLoanMapper,
+    val LoanRecord: LTLoanRecordMapper
+) {
+    suspend fun updateAssociatedLoanData(
+        transaction: Transaction?,
+        onBackgroundProcessingStart: suspend () -> Unit = {},
+        onBackgroundProcessingEnd: suspend () -> Unit = {},
+    ) {
+        computationThread {
+
+            if (transaction == null)
+                return@computationThread
+
+            if (transaction.loanId != null && transaction.loanRecordId == null) {
+                Loan.updateAssociatedLoan(
+                    transaction = transaction,
+                    onBackgroundProcessingStart = onBackgroundProcessingStart,
+                    onBackgroundProcessingEnd = onBackgroundProcessingEnd
+                )
+            } else if (transaction.loanId != null && transaction.loanRecordId != null) {
+                LoanRecord.updateAssociatedLoanRecord(
+                    transaction = transaction,
+                    onBackgroundProcessingStart = onBackgroundProcessingStart,
+                    onBackgroundProcessingEnd = onBackgroundProcessingEnd
+                )
+            }
+        }
+    }
+}
